requirements.txt
aiogram==3.22.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.3
aiofiles==23.2.1
aiohttp==3.9.5
python-dotenv==1.0.1
Flask==3.0.3
Flask-Limiter==3.12
gunicorn==22.0.0
Flask-WTF==1.2.1
requests==2.32.3
pyotp==2.9.0
qrcode==7.4.2